+++ packages/blockera-admin/js/dashboard.js
@@ -62,12 +62,14 @@ export const Dashboard = (): MixedElement => {
 		'blockera.admin.dashboard.pro.call.to.actions',
 		() => (
 			<Button
-				variant="secondary-on-hover"
-				icon={<Icon library={'ui'} icon={'crown'} iconSize={22} />}
-				text={__('Upgrade to Pro', 'blockera')}
+				variant="secondary"
 				href={window.blockeraUpgradeUrl}
 				target="_blank"
-			/>
+			>
+				<Icon library={'ui'} icon={'crown'} iconSize={22} />
+
+				{__('Upgrade to Pro', 'blockera')}
+			</Button>
 		)
 	);
 	const promotionComponent: MixedElement = applyFilters(
@@ -136,19 +138,23 @@ export const Dashboard = (): MixedElement => {
 
 	const profileComponent: MixedElement = applyFilters(
 		'blockera.admin.dashboard.profile.component',
-		<Promoter
-			heading={__('Upgrade to Blockera Pro', 'blockera')}
-			buttonText={__('Unlock All Features', 'blockera')}
-			style={{
-				width: '100%',
-				boxShadow: 'var(--card-box-shadow)',
-				borderRadius: 'var(--card-border-radius)',
-				padding: '25px 35px',
-				backgroundColor: 'var(--card-bg-color)',
-				marginTop: 'auto',
-				boxSizing: 'border-box',
-			}}
-		/>
+		<>
+			{!settings?.general?.disableProHints && (
+				<Promoter
+					heading={__('Upgrade to Blockera Pro', 'blockera')}
+					buttonText={__('Unlock All Features', 'blockera')}
+					style={{
+						width: '100%',
+						boxShadow: 'var(--card-box-shadow)',
+						borderRadius: 'var(--card-border-radius)',
+						padding: '25px 35px',
+						backgroundColor: 'var(--card-bg-color)',
+						marginTop: 'auto',
+						boxSizing: 'border-box',
+					}}
+				/>
+			)}
+		</>
 	);
 
 	const tabItems: Array<Object> = applyFilters(
@@ -201,25 +207,30 @@ export const Dashboard = (): MixedElement => {
 					version={`v${blockeraVersion}`}
 					name={__('Blockera', 'blockera')}
 				>
-					<div className={'blockera-settings-header-links'}>
+					<Flex
+						className={'blockera-settings-header-links'}
+						direction="row"
+						gap={10}
+						alignItems="center"
+					>
 						<ProCallToActions />
 
 						<Button
-							variant="tertiary-on-hover"
-							icon={
-								<Icon
-									library={'ui'}
-									icon={'changelog'}
-									iconSize={22}
-								/>
-							}
-							text={__('Changelog', 'blockera')}
+							variant="tertiary"
 							href={
 								'https://community.blockera.ai/changelog-9l8hbrv0?utm_source=blockera-admin&utm_medium=referral&utm_campaign=changelog-page&utm_content=cta-link'
 							}
 							target="_blank"
-						/>
-					</div>
+						>
+							<Icon
+								library={'ui'}
+								icon={'changelog'}
+								iconSize={22}
+							/>
+
+							{__('Changelog', 'blockera')}
+						</Button>
+					</Flex>
 				</Header>
 
 				<Tabs